2006-04-27  Sebastien Pouliot  <sebastien@ximian.com>
	* gdip.h, graphics.c: Fix prototypes for GdipDrawCurve3 and 
	GdipDrawCurve3I functions (float to int). The System.Drawing wrappers
	are correct. This is the first part of Duncan's patch for #76907.


svn path=/trunk/libgdiplus/; revision=60009

+++ src/gdip.h
@@ -1094,8 +1094,8 @@ GpStatus GdipDrawCurve (GpGraphics *graphics, GpPen *pen, GpPointF *points, int 
 GpStatus GdipDrawCurveI (GpGraphics *graphics, GpPen *pen, GpPoint *points, int count);
 GpStatus GdipDrawCurve2 (GpGraphics *graphics, GpPen *pen, GpPointF *points, int count, float tension);
 GpStatus GdipDrawCurve2I (GpGraphics *graphics, GpPen *pen, GpPoint *points, int count, float tension);
-GpStatus GdipDrawCurve3 (GpGraphics *graphics, GpPen *pen, GpPointF *points, int count, int offset, float numOfSegments, float tension);
-GpStatus GdipDrawCurve3I (GpGraphics *graphics, GpPen *pen, GpPoint *points, int count, int offset, float numOfSegments, float tension);
+GpStatus GdipDrawCurve3 (GpGraphics *graphics, GpPen *pen, GpPointF *points, int count, int offset, int numOfSegments, float tension);
+GpStatus GdipDrawCurve3I (GpGraphics *graphics, GpPen *pen, GpPoint *points, int count, int offset, int numOfSegments, float tension);
 GpStatus GdipDrawClosedCurve (GpGraphics *graphics, GpPen *pen, GpPointF *points, int count);
 GpStatus GdipDrawClosedCurveI (GpGraphics *graphics, GpPen *pen, GpPoint *points, int count);
 GpStatus GdipDrawClosedCurve2 (GpGraphics *graphics, GpPen *pen, GpPointF *points, int count, float tension);
